TACODEV-909: prevent error during wating for cluster
