+++ cluster-api-aws/.helmignore
@@ -23,3 +23,4 @@
 .vscode/
 artifacts/Dockerfile
 artifacts/helm
+artifacts/gh
