+++ cluster-api-aws/templates/configmap.yaml
@@ -7,11 +7,17 @@ data:
   generate_machine_pool.py: |
 {{ .Files.Get "artifacts/generate_machine_pool.py" | indent 4 }}
 
-  wait_and_initialize.py: |
-{{ .Files.Get "artifacts/wait_and_initialize.py" | indent 4 }}
+  node_label.py: |
+{{ .Files.Get "artifacts/node_label.py" | indent 4 }}
 
   mp.raw.yaml: |
 {{ tuple "mp/_mp.yaml" . | include "helper_template" | indent 4 }}
 
   argo-register.sh: |
 {{ .Files.Get "artifacts/argo-register.sh" | indent 4 }}
+
+  wait_for_kubeconfig.sh: |
+{{ .Files.Get "artifacts/wait_for_kubeconfig.sh" | indent 4 }}
+
+  wait_and_k8s_init.sh: |
+{{ .Files.Get "artifacts/wait_and_k8s_init.sh" | indent 4 }}
