+++ cluster-api-aws/templates/rbac.yaml
@@ -48,6 +48,22 @@ rules:
   - get
   - list
   - create
+- apiGroups:
+  - ''
+  resources:
+  - secrets
+  verbs: 
+  - get
+  - list
+  - watch
+- apiGroups:
+  - ''
+  resources:
+  - serviceaccount
+  verbs: 
+  - get
+  - list
+  - watch
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: RoleBinding
